Created Discount entity, implemented logic for removing product from sale, improved Product CRUD operations security

+++ src/main/java/com/serhiihurin/shop/online_shop/services/ProductServiceImpl.java
@@ -1,20 +1,25 @@
 package com.serhiihurin.shop.online_shop.services;
 
+import com.serhiihurin.shop.online_shop.dao.DiscountRepository;
 import com.serhiihurin.shop.online_shop.dao.ProductRepository;
+import com.serhiihurin.shop.online_shop.dao.ShopRepository;
 import com.serhiihurin.shop.online_shop.dto.ProductRequestDTO;
+import com.serhiihurin.shop.online_shop.entity.Discount;
 import com.serhiihurin.shop.online_shop.entity.Product;
-import com.serhiihurin.shop.online_shop.entity.Shop;
+import com.serhiihurin.shop.online_shop.entity.User;
 import com.serhiihurin.shop.online_shop.exception.ApiRequestException;
+import com.serhiihurin.shop.online_shop.exception.UnauthorizedAccessException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class ProductServiceImpl implements ProductService {
     private final ProductRepository productRepository;
+    private final ShopRepository shopRepository;
+    private final DiscountRepository discountRepository;
 
     @Override
     public List<Product> getAllProducts() {
@@ -23,17 +28,14 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public List<Product> getProductsByShopId(Long id) {
-        Optional<Product> optionalProductData = productRepository.findById(id);
-        if (optionalProductData.isEmpty()) {
-            throw new ApiRequestException("Could not find products from this shop");
-        }
-        return productRepository.getProductByShopId(id);
+        return productRepository.getProductsByShopId(id)
+                .orElseThrow(() -> new ApiRequestException("Could not find products from shop with ID: " + id));
     }
 
     @Override
     public Product getProduct(Long id) {
         return productRepository.findById(id)
-                .orElseThrow(() -> new ApiRequestException("Could not find product"));
+                .orElseThrow(() -> new ApiRequestException("Could not find product with ID: " + id));
     }
 
     @Override
@@ -43,12 +45,20 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public Product updateProduct(
+            User currentAuthenticatedUser,
             ProductRequestDTO productRequestDTO,
-            Shop shop,
             Product product
     ) {
-        if (shop != null) {
-            product.setShop(shop);
+        if (!product.getShop().getOwner().getId().equals(currentAuthenticatedUser.getId())) {
+            throw new UnauthorizedAccessException("Access denied. Wrong product ID");
+        }
+
+        if (productRequestDTO.getShopId() != null) {
+            product.setShop(
+                    shopRepository.findById(productRequestDTO.getShopId())
+                            .orElseThrow(() -> new ApiRequestException("Could not find shop with ID: " +
+                                    productRequestDTO.getShopId()))
+            );
         }
         if (productRequestDTO.getName() != null) {
             product.setName(productRequestDTO.getName());
@@ -59,23 +69,32 @@ public class ProductServiceImpl implements ProductService {
         if (productRequestDTO.getPrice() != null) {
             product.setPrice(productRequestDTO.getPrice());
         }
+        if (productRequestDTO.getAmount() != null) {
+            product.setAmount(productRequestDTO.getAmount());
+        }
 
         return productRepository.save(product);
     }
 
     @Override
-    public void increaseProductAmount(Product product, Integer amount) {
+    public void increaseProductAmount(User currentAuthenticatedUser, Product product, Integer amount) {
+        if (amount == null || amount < 0){
+            throw new ApiRequestException("Wrong amount parameter");
+        }
+        if (!product.getShop().getOwner().getId().equals(currentAuthenticatedUser.getId())) {
+            throw new UnauthorizedAccessException("Access denied. Wrong product ID");
+        }
         product.setAmount(product.getAmount() + amount);
         productRepository.save(product);
     }
 
     @Override
-    public void putProductOnSale(Long productId, int discountPercent) {
+    public void putProductOnSale(User currentAuthenticatedUser, Long productId, int discountPercent) {
         Product product = productRepository.findById(productId)
-                .orElseThrow(() -> new ApiRequestException("Could not find product"));
+                .orElseThrow(() -> new ApiRequestException("Could not find product with ID: " + productId));
 
-        if (product.isOnSale()) {
-            throw new ApiRequestException("The product is already on sale");
+        if (!product.getShop().getOwner().getId().equals(currentAuthenticatedUser.getId())) {
+            throw new UnauthorizedAccessException("Access denied. Wrong product ID");
         }
 
         if (discountPercent > 100 || discountPercent < 1) {
@@ -83,14 +102,49 @@ public class ProductServiceImpl implements ProductService {
                     "The value should be higher than 1 and lower than 100");
         }
 
+        if (product.isOnSale()) {
+            throw new ApiRequestException("The product is already on sale");
+        }
+
         product.setOnSale(true);
         product.setPrice(product.getPrice() - (product.getPrice() * discountPercent / 100));
 
+        discountRepository.save(
+                Discount.builder()
+                .product(product)
+                .discountPercent(discountPercent)
+                .build()
+        );
         productRepository.save(product);
     }
 
     @Override
-    public void deleteProduct(Long id) {
+    public void removeProductFromSale(User currentAuthenticatedUser, Long productId) {
+        Product product = productRepository.findById(productId)
+                .orElseThrow(() -> new ApiRequestException("Could not find product with ID: " + productId));
+
+        if (!product.getShop().getOwner().getId().equals(currentAuthenticatedUser.getId())) {
+            throw new UnauthorizedAccessException("Access denied. Wrong product ID");
+        }
+
+        Discount discount = discountRepository.findById(productId)
+                .orElseThrow(() -> new ApiRequestException("The product has no discount"));
+
+        product.setOnSale(false);
+        product.setPrice(product.getPrice() + (product.getPrice() * discount.getDiscountPercent() / 100));
+
+        discountRepository.deleteById(productId);
+        productRepository.save(product);
+    }
+
+
+    @Override
+    public void deleteProduct(User currentAuthenticatedUser, Long id) {
+        Product product = productRepository.findById(id)
+                .orElseThrow(() -> new ApiRequestException("Could not find product with ID: " + id));
+        if (!product.getShop().getOwner().getId().equals(currentAuthenticatedUser.getId())) {
+            throw new UnauthorizedAccessException("Access denied.");
+        }
         productRepository.deleteById(id);
     }
 }
